refactor nodegroup and cluster related functions out ov vpc_ipv4.go
