+++ pkg/cfn/builder/vpc_ipv4.go
@@ -373,197 +373,6 @@ type clusterSecurityGroup struct {
 	ClusterSharedNode *gfnt.Value
 }
 
-func (c *ClusterResourceSet) addResourcesForSecurityGroups(vpcID *gfnt.Value) *clusterSecurityGroup {
-	var refControlPlaneSG, refClusterSharedNodeSG *gfnt.Value
-
-	if c.spec.VPC.SecurityGroup == "" {
-		refControlPlaneSG = c.newResource(cfnControlPlaneSGResource, &gfnec2.SecurityGroup{
-			GroupDescription: gfnt.NewString("Communication between the control plane and worker nodegroups"),
-			VpcId:            vpcID,
-		})
-
-		if len(c.spec.VPC.ExtraCIDRs) > 0 {
-			for i, cidr := range c.spec.VPC.ExtraCIDRs {
-				c.newResource(fmt.Sprintf("IngressControlPlaneExtraCIDR%d", i), &gfnec2.SecurityGroupIngress{
-					GroupId:     refControlPlaneSG,
-					CidrIp:      gfnt.NewString(cidr),
-					Description: gfnt.NewString(fmt.Sprintf("Allow Extra CIDR %d (%s) to communicate to controlplane", i, cidr)),
-					IpProtocol:  gfnt.NewString("tcp"),
-					FromPort:    sgPortHTTPS,
-					ToPort:      sgPortHTTPS,
-				})
-			}
-		}
-	} else {
-		refControlPlaneSG = gfnt.NewString(c.spec.VPC.SecurityGroup)
-	}
-	c.securityGroups = []*gfnt.Value{refControlPlaneSG} // only this one SG is passed to EKS API, nodes are isolated
-
-	if c.spec.VPC.SharedNodeSecurityGroup == "" {
-		refClusterSharedNodeSG = c.newResource(cfnSharedNodeSGResource, &gfnec2.SecurityGroup{
-			GroupDescription: gfnt.NewString("Communication between all nodes in the cluster"),
-			VpcId:            vpcID,
-		})
-		c.newResource("IngressInterNodeGroupSG", &gfnec2.SecurityGroupIngress{
-			GroupId:               refClusterSharedNodeSG,
-			SourceSecurityGroupId: refClusterSharedNodeSG,
-			Description:           gfnt.NewString("Allow nodes to communicate with each other (all ports)"),
-			IpProtocol:            gfnt.NewString("-1"),
-			FromPort:              sgPortZero,
-			ToPort:                sgMaxNodePort,
-		})
-	} else {
-		refClusterSharedNodeSG = gfnt.NewString(c.spec.VPC.SharedNodeSecurityGroup)
-	}
-
-	if c.supportsManagedNodes && api.IsEnabled(c.spec.VPC.ManageSharedNodeSecurityGroupRules) {
-		// To enable communication between both managed and unmanaged nodegroups, this allows ingress traffic from
-		// the default cluster security group ID that EKS creates by default
-		// EKS attaches this to Managed Nodegroups by default, but we need to handle this for unmanaged nodegroups
-		c.newResource(cfnIngressClusterToNodeSGResource, &gfnec2.SecurityGroupIngress{
-			GroupId:               refClusterSharedNodeSG,
-			SourceSecurityGroupId: gfnt.MakeFnGetAttString("ControlPlane", outputs.ClusterDefaultSecurityGroup),
-			Description:           gfnt.NewString("Allow managed and unmanaged nodes to communicate with each other (all ports)"),
-			IpProtocol:            gfnt.NewString("-1"),
-			FromPort:              sgPortZero,
-			ToPort:                sgMaxNodePort,
-		})
-		c.newResource("IngressNodeToDefaultClusterSG", &gfnec2.SecurityGroupIngress{
-			GroupId:               gfnt.MakeFnGetAttString("ControlPlane", outputs.ClusterDefaultSecurityGroup),
-			SourceSecurityGroupId: refClusterSharedNodeSG,
-			Description:           gfnt.NewString("Allow unmanaged nodes to communicate with control plane (all ports)"),
-			IpProtocol:            gfnt.NewString("-1"),
-			FromPort:              sgPortZero,
-			ToPort:                sgMaxNodePort,
-		})
-	}
-
-	if c.spec.VPC == nil {
-		c.spec.VPC = &api.ClusterVPC{}
-	}
-	c.rs.defineOutput(outputs.ClusterSecurityGroup, refControlPlaneSG, true, func(v string) error {
-		c.spec.VPC.SecurityGroup = v
-		return nil
-	})
-	c.rs.defineOutput(outputs.ClusterSharedNodeSecurityGroup, refClusterSharedNodeSG, true, func(v string) error {
-		c.spec.VPC.SharedNodeSecurityGroup = v
-		return nil
-	})
-
-	return &clusterSecurityGroup{
-		ControlPlane:      refControlPlaneSG,
-		ClusterSharedNode: refClusterSharedNodeSG,
-	}
-}
-
-// TODO move this
-func (rs *resourceSet) addEFASecurityGroup(vpcID *gfnt.Value, clusterName, desc string) *gfnt.Value {
-	efaSG := rs.newResource("EFASG", &gfnec2.SecurityGroup{
-		VpcId:            vpcID,
-		GroupDescription: gfnt.NewString("EFA-enabled security group"),
-		Tags: []gfncfn.Tag{{
-			Key:   gfnt.NewString("kubernetes.io/cluster/" + clusterName),
-			Value: gfnt.NewString("owned"),
-		}},
-	})
-	rs.newResource("EFAIngressSelf", &gfnec2.SecurityGroupIngress{
-		GroupId:               efaSG,
-		SourceSecurityGroupId: efaSG,
-		Description:           gfnt.NewString("Allow " + desc + " to communicate to itself (EFA-enabled)"),
-		IpProtocol:            gfnt.NewString("-1"),
-	})
-	rs.newResource("EFAEgressSelf", &gfnec2.SecurityGroupEgress{
-		GroupId:                    efaSG,
-		DestinationSecurityGroupId: efaSG,
-		Description:                gfnt.NewString("Allow " + desc + " to communicate to itself (EFA-enabled)"),
-		IpProtocol:                 gfnt.NewString("-1"),
-	})
-
-	return efaSG
-}
-
-// TODO move this
-func (n *NodeGroupResourceSet) addResourcesForSecurityGroups() {
-	for _, id := range n.spec.SecurityGroups.AttachIDs {
-		n.securityGroups = append(n.securityGroups, gfnt.NewString(id))
-	}
-
-	if api.IsEnabled(n.spec.SecurityGroups.WithShared) {
-		n.securityGroups = append(n.securityGroups, n.vpcImporter.SharedNodeSecurityGroup())
-	}
-
-	if api.IsDisabled(n.spec.SecurityGroups.WithLocal) {
-		return
-	}
-
-	desc := "worker nodes in group " + n.spec.Name
-	vpcID := n.vpcImporter.VPC()
-	refControlPlaneSG := n.vpcImporter.ControlPlaneSecurityGroup()
-
-	refNodeGroupLocalSG := n.newResource("SG", &gfnec2.SecurityGroup{
-		VpcId:            vpcID,
-		GroupDescription: gfnt.NewString("Communication between the control plane and " + desc),
-		Tags: []gfncfn.Tag{{
-			Key:   gfnt.NewString("kubernetes.io/cluster/" + n.clusterSpec.Metadata.Name),
-			Value: gfnt.NewString("owned"),
-		}},
-		SecurityGroupIngress: makeNodeIngressRules(n.spec.NodeGroupBase, refControlPlaneSG, n.clusterSpec.VPC.CIDR.String(), desc),
-	})
-
-	n.securityGroups = append(n.securityGroups, refNodeGroupLocalSG)
-
-	if api.IsEnabled(n.spec.EFAEnabled) {
-		efaSG := n.rs.addEFASecurityGroup(vpcID, n.clusterSpec.Metadata.Name, desc)
-		n.securityGroups = append(n.securityGroups, efaSG)
-	}
-
-	n.newResource("EgressInterCluster", &gfnec2.SecurityGroupEgress{
-		GroupId:                    refControlPlaneSG,
-		DestinationSecurityGroupId: refNodeGroupLocalSG,
-		Description:                gfnt.NewString("Allow control plane to communicate with " + desc + " (kubelet and workload TCP ports)"),
-		IpProtocol:                 sgProtoTCP,
-		FromPort:                   sgMinNodePort,
-		ToPort:                     sgMaxNodePort,
-	})
-	n.newResource("EgressInterClusterAPI", &gfnec2.SecurityGroupEgress{
-		GroupId:                    refControlPlaneSG,
-		DestinationSecurityGroupId: refNodeGroupLocalSG,
-		Description:                gfnt.NewString("Allow control plane to communicate with " + desc + " (workloads using HTTPS port, commonly used with extension API servers)"),
-		IpProtocol:                 sgProtoTCP,
-		FromPort:                   sgPortHTTPS,
-		ToPort:                     sgPortHTTPS,
-	})
-	n.newResource("IngressInterClusterCP", &gfnec2.SecurityGroupIngress{
-		GroupId:               refControlPlaneSG,
-		SourceSecurityGroupId: refNodeGroupLocalSG,
-		Description:           gfnt.NewString("Allow control plane to receive API requests from " + desc),
-		IpProtocol:            sgProtoTCP,
-		FromPort:              sgPortHTTPS,
-		ToPort:                sgPortHTTPS,
-	})
-}
-
-func makeNodeIngressRules(ng *api.NodeGroupBase, controlPlaneSG *gfnt.Value, vpcCIDR, description string) []gfnec2.SecurityGroup_Ingress {
-	ingressRules := []gfnec2.SecurityGroup_Ingress{
-		{
-			SourceSecurityGroupId: controlPlaneSG,
-			Description:           gfnt.NewString(fmt.Sprintf("[IngressInterCluster] Allow %s to communicate with control plane (kubelet and workload TCP ports)", description)),
-			IpProtocol:            sgProtoTCP,
-			FromPort:              sgMinNodePort,
-			ToPort:                sgMaxNodePort,
-		},
-		{
-			SourceSecurityGroupId: controlPlaneSG,
-			Description:           gfnt.NewString(fmt.Sprintf("[IngressInterClusterAPI] Allow %s to communicate with control plane (workloads using HTTPS port, commonly used with extension API servers)", description)),
-			IpProtocol:            sgProtoTCP,
-			FromPort:              sgPortHTTPS,
-			ToPort:                sgPortHTTPS,
-		},
-	}
-
-	return append(ingressRules, makeSSHIngressRules(ng, vpcCIDR, description)...)
-}
-
 func (v *IPv4VPCResourceSet) haNAT() {
 	for _, az := range v.clusterConfig.AvailabilityZones {
 		alphanumericUpperAZ := formatAZ(az)
